requirements.txt
absl-py==0.3.0
astor==0.7.1
backcall==0.1.0
bleach==1.5.0
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
gast==0.2.0
GDAL==2.2.2
grpcio==1.14.1
html5lib==0.9999999
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
numpy==1.15.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toolz==0.9.0
tqdm==4.24.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
